feat(tui): integrate BorderPulse into InferBox
Add pulse_intensity field to InferBox for animated border effects during
Running state. The border brightens up to 30% based on pulse intensity.

Changes:
- Add pulse_intensity: f32 field (default 0.0)
- Add with_pulse_intensity() builder method with clamping
- Update render() to use border_color_with_pulse()
- Add 3 tests for pulse functionality

+++ tools/nika/src/tui/widgets/matrix_rain.rs
@@ -735,7 +735,7 @@ mod tests {
             let rain = MatrixRain::new()
                 .frame(frame)
                 .with_nika_pattern(true)
-                .explosion_frame(frame as u8)
+                .explosion_frame(frame)
                 .opacity(1.0);
             let mut buf = Buffer::empty(area);
             rain.render(area, &mut buf);
